// src/domain/enums/app-expection-codes.enum.ts
enum AppExceptionCodesEnum {
  CONTINUE = 'continue',
  SWITCHING_PROTOCOLS = 'switching_protocols',
  OK = 'ok',
  CREATED = 'created',
  ACCEPTED = 'accepted',
  NON_AUTHORITATIVE_INFORMATION = 'non_authoritative_information',
  NO_CONTENT = 'no_content',
  RESET_CONTENT = 'reset_content',
  PARTIAL_CONTENT = 'partial_content',
  MULTIPLE_CHOICES = 'multiple_choices',
  MOVED_PERMANENTLY = 'moved_permanently',
  FOUND = 'found',
  SEE_OTHER = 'see_other',
  NOT_MODIFIED = 'not_modified',
  USE_PROXY = 'use_proxy',
  TEMPORARY_REDIRECT = 'temporary_redirect',
  BAD_REQUEST = 'bad_request',
  UNAUTHORIZED = 'unauthorized',
  PAYMENT_REQUIRED = 'payment_required',
  FORBIDDEN = 'forbidden',
  NOT_FOUND = 'not_found',
  METHOD_NOT_ALLOWED = 'method_not_allowed',
  NOT_ACCEPTABLE = 'not_acceptable',
  PROXY_AUTHENTICATION_REQUIRED = 'proxy_authentication_required',
  REQUEST_TIMEOUT = 'request_timeout',
  CONFLICT = 'conflict',
  GONE = 'gone',
  LENGTH_REQUIRED = 'length_required',
  PRECONDITION_FAILED = 'precondition_failed',
  REQUEST_ENTITY_TOO_LARGE = 'request_entity_too_large',
  REQUEST_URI_TOO_LONG = 'request_uri_too_long',
  UNSUPPORTED_MEDIA_TYPE = 'unsupported_media_type',
  REQUESTED_RANGE_NOT_SATISFIABLE = 'requested_range_not_satisfiable',
  EXPECTATION_FAILED = 'expectation_failed',
  UNPROCESSABLE_ENTITY = 'unprocessable_entity',
  TOO_MANY_REQUESTS = 'too_many_requests',
  INTERNAL_SERVER_ERROR = 'internal_server_error',
  NOT_IMPLEMENTED = 'not_implemented',
  BAD_GATEWAY = 'bad_gateway',
  SERVICE_UNAVAILABLE = 'service_unavailable',
  GATEWAY_TIMEOUT = 'gateway_timeout',
  VERSION_NOT_SUPPORTED = 'version_not_supported',
}

export default AppExceptionCodesEnum;